Refactoring
- Add SharedPresentation class for methods in GUIUtil which are no
utility methods.
- Move restoreSeedWords to SharedPresentation
